fix(tarko): remove auto-scroll logic, keep manual scroll-to-bottom
